Executor support for NodeAdd and NodeDelete
Summary of features:

* Configuration file support
* Executor: SSH, execute commands on actual systems
* Executor: Mock, no commands executed.
* Functional tests against a running server and GlusterFS cluster

This commit contains the support for NodeAdd and NodeDelete
to do a PeerProbe and a PeerDetach, respectively.  The work
done to do the actual work is done through the Executor interface.
There are two implentatatinos of the Executor interface.  One is
the 'ssh' implementation, which executes the gluster commands
on an actual system.  The second is a 'mock' version, which allows
the use of the Heketi server without actually sending any commands,
which is necessary for unit tests.

Also added support for configuration files and flags for Heketi. The
configuration file can now pass a lot of values to Heketi, like server
port, and ssh information.  The configuration file is in JSON format,
which makes it really easy to read from Go.

Lastly, a functional test was started to execute against a running
Heketi server on localhost.  The functional tests assume that the
Vagrant setup create by vagrant-heketi is setup and running.  To run
the functional tests you must type:

    $ go test -tags 'functionaltests'

Closes #119
Closes #118
Closes #86
Closes #19

Signed-off-by: Luis Pabón <lpabon@redhat.com>