Functional Tests
Functional tests now use the Heketi client API.  There are two
types of tests.  One is the 'small' functional test, enough
for testing on a system (like a laptop) with at least 8GB RAM.
The second is for a 'large' functional test.  This requires
a single system with at least 128GB of RAM.

Also Hello() has been added to the Heketi Client API.

Signed-off-by: Luis Pabón <lpabon@redhat.com>