Heketi CLI
heketi-cli can be used to communicate with the Heketi server.
The cli itself is not meant to be used for regular user
consumption, instead it is meant as a tool to be used by
developers and administrators.  Services like Kubernetes,
OpenStack Manila, OpenShift and others are the primary
consumers of the REST API.

Signed-off-by: Luis Pabón <lpabon@redhat.com>