Node and device state support
New APIs are now available which allow callers to set a node
or a device offline and online again.  If a node is placed
offline, then all of its disks are removed from the ring. If
a device is moved offline, then only that device is removed
from the ring.

New State API is available to accomplish this:
  POST /devices/<id>/state
  POST /nodes/<id>/state

The normal XInfo API calls can be used to get state information.

This patch also includes the following changes:

* All API structures moved to pkg/glusterfs/api
* Go and Python Client API support for state
* Heketi-cli support for state

Signed-off-by: Luis Pabón <lpabon@redhat.com>