app: Correctly handle stale registrations
Heketi uses the concept of registrations to manage conflict
resolution.  For example, a device `/dev/sda` cannot be added
twice to the same node.  In the same manner a node with the
same hostnames cannot be added twice.

To resolve conflicts, Heketi _Registers_ the conflicting string
as a _key_ in the database.  If for any reason the device or
node was unable to be added, then Heketi has methods to remove
that registration.

The issue arises when Heketi is adding a device, then crashes
for any number of reasons.  At this point, the registered key
is still in the database for conflict resolution, but the entry
does not exist.

This patch checks to make sure that a registered name has an
entry in the database, and it is not a stale key describing
and unknown device or node.

Closes #361

Signed-off-by: Luis Pabón <lpabon@redhat.com>