Kubernetes DaemonSet Support
This change allows Heketi to discover and manage GlusterFS containers which
have been deployed by using a DaemonSet.  The GlusterFS deploy object
has been changed to a GlusterFS daemonset which will need to be installed
by the administrator.  When the administrator wants the DaemonSet to
deploy GlusterFS to a specific node, the administrator can then label
that node with `storagenode=glusterfs`.  This is setup in the DaemonSet
as the `nodeSelector`, which will instruct Kubernetes to deploy the
glusterfs container to the specified node.

When a node is added to Heketi to be managed, just as before, the management
host is the name of the node as displayed in `kubectl get nodes`.  When
Heketi needs to execute a command in a GlusterFS container which has been
deployed by a daemonset, it will fetch all pods in the namespace which
match the label `glusterfs-node` (see kubeexec.go:36).  It then receives
a list of all the pods which match this label.  Heketi goes through each
pod spec looking for the one which matches the management host name.  Once
the pod has been found, it uses the pod name to execute commands on that
container.

Signed-off-by: Luis Pabón <luis.pabon@coreos.com>