Topology Info, Node Info: "ID not found" error when running concurrently with volume creation.
The brick entries which is being updated during volume create causes inconsitent state. In the state, It returns "Id not found" actually for bricks Id. This fix will make node info and topology info not to depend on the Bricks. In actual senario it is not using the bricks anyways.

Signed-off-by: Mohamed Ashiq Liyazudeen <mliyazud@redhat.com>