tests: add TestVolumeCreateVsTopologyInfo()
Test for https://github.com/heketi/heketi/issues/382 :
A TopologyInfo request running concurrently to a VolumeCreate
failed with "Id not found" due to unsaved brick info.
Run a couple of these concurrently to increase the chance
for hitting the race.

When you revert the patch
"app: save bricks in allocBricks() to keep the db consistent",
then this test fails.

Closes #382

Signed-off-by: Michael Adam <obnox@redhat.com>