Automatic Kubernetes configuration in Heketi
With Kubernetes 1.5 we can now use higher level APIs to be able to automatically
setup REST RPC configuration with Kubernetes. There is no longer
a need to ask Namespace, API host, or Secret from the administrator with this
change.  The configuration also sets up an HTTPS encrypted connection to the
API host, which previous versions did not.

I tried first to use go-client for Kubernetes, but it does not provide
ability to setup and execute commands on containers.  Instead, this changes
enables Heketi to use Kubernetes 1.5 APIs to automatically setup configuration
with the API host.

Godeps had to be updated from Kubernetes 1.2.alpha to 1.5.beta.2.

This change also makes it necessary to deploy Heketi *in* Kubernetes
when using kubexec.

Closes #615

Signed-off-by: Luis Pabón <luis.pabon@coreos.com>