provide flag to specify heketi image in copy job
setup-openshift-heketi-storage command creates a Gluster volume to
store heketi db. The image that is used to create the Gluster volume is
hard coded to be heketi/heketi:dev. It is useful to have a cli option
to override this for a user who might want to use a different version
of the image like heketi:ci or even a different image of heketi from a
different registry.

Usage:
heketi-cli setup-openshift-heketi-storage --image "IMAGE_NAME"

image flag is optional and the default is still "heketi/heketi:dev"

Signed-off-by: Raghavendra Talur <rtalur@redhat.com>