Compress database when storing in a K8S Secret
With this change, the database is compressed before storing in
a Kuberntes Secret, which currently has a maximum size of 1Mi.
Because of this limit, it has been found that the maximum number
of devices allowed is 12000. This creates a db which is around 30Mi,
but is compressed down to around 900Ki.

Closes #697

Signed-off-by: Luis Pabón <luis.pabon@coreos.com>