save and reuse volume gid in volume entry
Earlier, VolumeEntry.gidRequested wasn't being stored in db because
Marshal/UnMarshal methods rely on reflection to get members of a struct
to be stored in db. gidRequested being a unexported member, it wasn't
being stored. VolumeEntry.Info.Gid was not being set to gidRequested,
hence it was always found to be zero.

This patch deletes gidRequested from VolumeEntry(as it was never
exported, it is safe to do so) and uses VolumeEntry.Info.Gid in all the
further operations.

Signed-off-by: Raghavendra Talur <rtalur@redhat.com>