apps: glusterfs: test: Use standard testing library
For reasons unknown to the patch author, a special "heketi/tests"
library wrapper was used instead of the standard testing semantics. This
makes it difficult for new users to read the code, and hard to trace
things properly.

As well, this code can now "return" and end a test early without
resorting to a panic.

While the current code is now longer, this is due to the fact that this
"unit" test is more of an "integration test" and tests all sorts of
things constantly, even if many probably won't fail and which don't add
much value to the test. Simplifying that can be done in future patches.

This new style of testing (as seen in the golang stdlib) should be used
for all patches going forward.