apps: glusterfs: Use AddCluster method
This now fits with the best API design of requiring users to add
clusters before they add devices. Previously it was done silently and
implicitly. Now you should do it explicitly if you can, with the
exception that we still do it for you for now in the simple allocator.

NOTE: this is based on methodology that obnoxxx requested, as my variant
of this patch was different.