utils: eliminate usage of stringset type
Replace and remove the stringset type from the code and corresponding
tests. It is no longer needed due to the loop not producing duplicate
host names. It aslo had a couple of issues IMO: It was only being used
in this one place and it used an append-and-sort method to add items
to the set.

Signed-off-by: John Mulligan <jmulligan@redhat.com>