extras: fix fromsource Dockerfile for current fedora (fix #923)
It appears that in f27 make has become a dependency of dnf.
The RUN line in the dockerfile was trying to remove make causing
dnf to exit nonzero and failing the build.

I broke up some of the multi-action lines so that its clearer
what steps are being taken. I also kept the instruction to
install make but combined it with the existing dnf install
line. Finally, we remove the -q from the dnf calls
as it was also supressing error output making the failure
very mysterious.

Signed-off-by: John Mulligan <jmulligan@redhat.com>