apps: create a logging rest api to get & set log level
This change implements the server side of issue #223.
We extend the api so that GETs and POSTs to the /internal/logging
endpoint can view and modify the logging level.

The api is formatted in such a way that multiple loggers can
be supported without too much fuss in the future. Currently,
it can only view or modify the logger internal to the glusterfs
app and does not have an impact on any of the other loggers
within heketi.

Signed-off-by: John Mulligan <jmulligan@redhat.com>