doc: fix steps and generally improve kubernetes install doc
This improves the issues reported in #931

The install doc for kubernetes was out of date/wrong in various
regards. This change improves the document so that it is more
correct, more readable and has better wording.

One should now be able to run through the steps and have a
working example/test environment using Heketi & Gluster under
Kubrnetes.

Signed-off-by: John Mulligan <jmulligan@redhat.com>