attempt to workaround tox in travis issue
Something there is that doesn't love a python3.5,
at least in tox under travis.

Travis continues to give our python unit tests problems when trying
to use python3. This may be due to the fact we are a Go project
that is also trying to do some python. Or I could just be a noob
who still doesn't understand things. :-) This attempts to concentrate
the logic in our test script and check the environment before
handing off to tox.

I don't know yet if this will fix the issue, but
I have to push to find out.

Signed-off-by: John Mulligan <jmulligan@redhat.com>