add db import and export functionality
This provides a mechanism to dump the contents of a db file into JSON
and the reverse. As the db contains 7 different buckets(as of today),
we create an encapsulating struct to bring them together.

It adds a new REST URL "/db/dump" that provides the db state as json
over http GET.
It also provides two subcommands to heketi binary that can be used to
export db state to a json file and import contents of a json file into
a new heketi db.

As explained above, import feature can only be used to create a new db
and not to add data to existing one.

If the db file is in use by another heketi process then the subcommands
will fail.

Signed-off-by: Raghavendra Talur <rtalur@redhat.com>