apps: prepare TestClusterDelete for lazy loading allocator
In the current test code, if the allocator is lazily loaded it
fails to initialize. This is because the allocator setup code
expects a well-formed db and the test was populating the db
in a minimal way. This was fine when the allocater was eagerly
loaded before the db population (nothing to fail on).

Change the test to more fully pouluate the db (at least
enough to have all the objects the allocator needs). I also
take the opportunity to make some the naming more specific
so we don't get lost in piles of generic "entry"s.

Signed-off-by: John Mulligan <jmulligan@redhat.com>