utils: add a non-random source of randomness for testing
Add a very simple non-random source of "randomness" for the UUID
generator. This can & will be used by future unit test code such that
it can compare object ids more simply (at least when the test is not
heavily threaded). This can also be used when debugging existing
failing unit tests (which is why I'm adding this now).

Add tests for the linear ids and for random source pluggability.

Signed-off-by: John Mulligan <jmulligan@redhat.com>