apps: make brick allocation function db access read-only
Make our new brick allocation function access the db in a read-only
mode. This change lifts the db writes back up to the caller of our
allocation function. To make things work without intermediate writes
to the db we add a `devcache` mapping to preserve space allocations
to the device objects we are using between trips around the top-level
loop.

Signed-off-by: John Mulligan <jmulligan@redhat.com>