utils: support generating "short ids", useful for tests
There are a bunch of places in the test code that generate a uuid and
then throw away a big chunk of it in an effort to create a semi-unique
id for some test-generated item. This change extends the API in uuid.go
to provide a ShortID function which returns a semi-unique id between 2-32
characters.

Signed-off-by: John Mulligan <jmulligan@redhat.com>