[RFC] apps: add a lock around brick/device allocation
Due to how the DB access was split from writing the db, concurrent
calls to allocateBricks could return Device objects that would
stomp on each other when saved to the DB. This change adds a lock
around the device/brick allocation and in order to keep concurrent
volume create requests functioning.

I've marked this particular change as RFC in order to get some feedback
on if the team feels like this is a good direction. I think the only
alternative approach would be to try and do the entire allocation
within a single db write transaction. I'm honestly not sure what would
be better but since this is less code ATM this is what I'm putting
in my PR for now.

Signed-off-by: John Mulligan <jmulligan@redhat.com>