tests: always run the functional tests with the verbose switch
The 'go test' -v switch is needed in order for the test suite to write
out what test cases it is running. This is needed for debugging sessions
to find out what test case failed and what case the corresponding server
is coming from.

Compared to the output the heketi server itself outputs the extra amount
of text is marginal, so why not leave it on all the time.

Signed-off-by: John Mulligan <jmulligan@redhat.com>