apps: only assign bricks to volume on complete success
During the brick allocation process multiple passes are made over
the brick allocation parameters. If bricks are instantiated and
added to the volume but the overall pass fails we can "leak" brick
ids in volume objects but these bricks never get created, leading to
future issues. This change makes sure only the final set of bricks
is added to the volume object.

Signed-off-by: John Mulligan <jmulligan@redhat.com>