apps: predetermine brick path on brick entries
Previously the executor code was changed to accept pre-determined
brick paths. This change pushes the predetermination up even earlier,
out of brick create, into brick instantiation. Now bricks entries are "born"
with paths that are known and these can be used in the brick creation
functions. Note that any time the brick ID is changed the path must be
kept in sync via the UpdatePath function - this is done automatically
when making a new brick object or calling SetId.

Signed-off-by: John Mulligan <jmulligan@redhat.com>