tests: do not fetch go binaries if not needed
Change the functional tests so they support only fetching go binaries
if needed. Use the HEKETI_TEST_SYSTEM_GO variable to allow the user
to tweak this behavior. The value of this variable can be yes, no, or
auto. Auto will only fetch go binaries if the version number is older
than 1.8.3, and is the default behavior.

Signed-off-by: John Mulligan <jmulligan@redhat.com>