tests: make logging in unit tests consistent
Some tests were disabling logs and some were not. Based on the
discussion in #792, changing everywhere to default logging.

As per https://github.com/golang/go/issues/21461, the logs won't
pollute the output if all tests pass.

If a test fails then all logs from the package are shown. This is
something which can be fixed later. It would be useful to limit logs of
that particular test instead of package.

Signed-off-by: Raghavendra Talur <rtalur@redhat.com>