apps: add a general operation retry mechanism
Build a retry mechanism into RunOperation and AsyncHttpOperation,
such that if the operation fails with an error of type
OperationRetryError and the Operation provides a greater-than-zero
MaxRetries value the operation will be re-run until the
max retries are exceeded or all operation phases succeed.

The operation Rollback method _must_ set all inputs to the operation
back to a state where Build can be safely executed again.

Signed-off-by: John Mulligan <jmulligan@redhat.com>