apps:glusterfs: store the allocator in the app again.
Previously, the allocator was not stored in the app
to prevent stale data as loaded from the db. Now that
the data is always loaded freshly in GetNodes(),
the stored allocator is not really a cache any more
but basically reflects the chosen allocator methods.
So at app start, we instantiate and store the configured
allocator again.

Signed-off-by: Michael Adam <obnox@redhat.com>