apps: use intermediate brick & device set types
Now that we have brick and device set types use them rather
than a flat list of bricks and devices.
The downside to this change is the nested looping required
to iterate over individual bricks and devices. I thought about
adding a helper method for this but it then seemed like
overkill for now. We could always add one later.

Signed-off-by: John Mulligan <jmulligan@redhat.com>