apps: support predicate func so callers can filter devices during alloc
Instead of needing to predict all the conditions a caller of our
allocation functions will need we add support for a predicate function
which can be used to reject device entries from a brick set.
A nil predicate function is treated the same as a predicate function
that always returns true.

Signed-off-by: John Mulligan <jmulligan@redhat.com>