apps: reduce the size of arbiter bricks relative to data bricks
The arbiter bricks (and only the arbiter bricks) are to be sized
smaller than the data bricks because they only hold metadata
corresponding to the metadata and data on the data bricks.

The patch adds a shim object partially because I couldn't stand
the thought of adding yet another argument to some of these
functions. But it also provides for a nice conceptual model
I think. Let me know if you disagree.

Currently the discount factor is hard coded but we may
eventually want it configurable on a per-volume basis.

Math provided by: Raghavendra Talur <rtalur@redhat.com>

Signed-off-by: John Mulligan <jmulligan@redhat.com>