server: consolidate glusterfs app config reading
Previously, the glusterfs app expected a file handle and
read the config from that. This changes the app so that
the config is passed in and the config file processing is
now all handled by the server/config package.

Now that the NewApp function takes a config object and not
an open file handle we need to update the tests correspondingly.
Some tests are pretty pointless in this new world and are
removed.

Signed-off-by: John Mulligan <jmulligan@redhat.com>