tests: don't write files when running functional test compile check
We recently added a script to compile the functional test go code as
a faster (and more frequently run) sanity check. However, the
go command line tool defaults to writing a <pkg>.test binary when
asked to compile test code. We tell the go tool to write the result
to /dev/null so we don't generate unwanted artifacts any more.

Signed-off-by: John Mulligan <jmulligan@redhat.com>