executor: revert add-brick if rebalance fails
If rebalance fails, earlier we were considering the whole add-brick
operation to be a failure. However, this would clean the bricks without
reverting the state of volume on gluster.

With this change, if rebalance on a volume fails after add-brick, we try
to revert the add-brick operation. If successful to revert, we can go
ahead with a proper Rollback. If not successful, we fake a successful
add-brick; this ensures gluster and heketi are in sync and gives admin a
chance to initiate rebalance manually.

Co-authored-by: Madhu Rajanna <mrajanna@redhat.com>
Signed-off-by: Raghavendra Talur <rtalur@redhat.com>