apps: add a node health monitor & cache subsystem
This change adds a new go file that defines a new type that
can monitor the health of nodes and cache the results
in memory for use by other parts of Heketi. A node is defined
as up/healthy if it can be contacted and is running gluster.

The type's Monitor method will start a background goroutine that
periodically checks in on the nodes. The time between monitor start
and initial check and between periodic checks are passed to
the type at initialization.

Co-authored-by: Raghavendra Talur <rtalur@redhat.com>
Co-authored-by: Madhu Rajanna <mrajanna@redhat.com>
Signed-off-by: Madhu Rajanna <mrajanna@redhat.com>
Signed-off-by: John Mulligan <jmulligan@redhat.com>