cmdexec: cleanly destroy bricks w/ thin-pool
After a volume has been cloned, multiple bricks use the same thin-pool
LV. When a brick is destroyed, the thin-pool LV can only be removed if
there are no consumers anymore. Bricks that are not the only user of the
thin-pool LV can still be destroyed, but the thin-pool LV needs to
remain.

In addition to the above, GlusterD manages the mounts of bricks that
have been created when a volume was cloned. These bricks are not listed
in /etc/fstab, and can therefore not be removed from there either.

Signed-off-by: Niels de Vos <ndevos@redhat.com>