volume-clone: account for space reclaimed when bricks are destroyed
Bricks that have been created as part of a volume-clone operation share
the thin-pool of the orignal bricks. The cloned bricks live on the same
devices as the original ones.

This causes the information from the BrickEntry to become unsuitable for
calculating the reclaimed space when a brick is destroyed. The details
in the BrickEntry do not capture anything about the shared usage of the
thin-pool.

In order to calculate the space that has been free'd when the brick is
destroyed by the executor, the Brick.Destroy() function now returns
(bool, error), where the bool indicates if the thin-pool has been
free'd. In case it is not free'd, the Destroy() of the Brick did not
free any space on the device.

The VolumeDeleteOperation now gathers the return value from the
executors Brick.Destroy(), and uses that to calculate the space that the
brick had allocated (including alignment etc.). With this, the device
can get updated with the space that has been reclaimed.

Signed-off-by: Niels de Vos <ndevos@redhat.com>