apps: enable the node health monitor by default in server binary
This change removes the ability to enable or disable the node health
monitor from the heketi config file. Instead, we provide a package
level variable that the glusterfs app uses to determine if the
monitor should be used or not. By default this variable is false thus
all unit tests and other non-server code will not run the monitor.

The server binary does set this variable to true by default thus
enabling the monitor for the heketi server. For debugging purposes
this can be disabled (in the server only) by setting
`HEKETI_DISABLE_HEALTH_MONITOR=true` in the environment.

Signed-off-by: John Mulligan <jmulligan@redhat.com>