apps: make the volume create rollback better match reality
If the rollback function can not determine if gluster has
removed the in-progress volume or not it should not treat
ignorance as success. Add an attempt to delete the volume
on the gluster side and use its result to determine if
the volume has been removed or not.
If not, the pending volume will remain in the db, hopefully
to be cleaned up later.

Signed-off-by: John Mulligan <jmulligan@redhat.com>