apps: make device check callbacks mandatory for arbiter placer
In order to better split what logic lives in the glusterfs app
vs the planned placer package we need keep the tag checking logic
out of the arbiter placer proper. This means no more default
implementation of arbiter and data device checks. We require
the caller to always provide the functions.

Signed-off-by: John Mulligan <jmulligan@redhat.com>