tests: allow each test.go file to have heketi config
With so many new features being added to heketi that have config
options to be specified in heketi.json, it is useful to start heketi
server with a different config for each test.

To keep things simple, we expect the config file to be named same as the
test file but with "json" as extension instead of "go". In case where
such a file isn't found, we use the default config file heketi.json in
config dir of test dir.

Signed-off-by: Raghavendra Talur <rtalur@redhat.com>