functional tests: lower size of temporary volumes
The TestHeketiSmokeTest was failing intermittently due to the
way devices are selected for bricks (randomization in go maps, etc).
This adds a unit test to prove the values for the functional
test and updates the functional test with a lower value.
If we ever revisit how the ring logic works we could possibly
re-up the value.

Signed-off-by: John Mulligan <jmulligan@redhat.com>