utils: improve error reporting if service does not provide a message
On occasion the heketi-cli returned an empty `Error: ` to the caller.
This is obviously suboptimal, as there is no hint to what could have
gone wrong. The Heketi service itself provides error messages for any
issues it hits. If there are no error messages received by the
heketi-cli, the problem is somewhere else.

It seems that a misconfigured router/proxy in Kubernetes or OpenShift
can cause empty error messages. These routers/proxies will provide a
HTTP status code, so let's display that if there is no message.

Bug: https://bugzilla.redhat.com/1432340
Signed-off-by: Niels de Vos <ndevos@redhat.com>