arbiter: fix volume creation with too big average file size
The arbiter brick size for a brick set is calculated as
(dataBrickSize)/(averageFileSize). If the data brick is
rather small, and the averageFileSize is rather big, this
can get smaller than the minimum filesystem size of 16MiB
that xfs supports (with 4KiB blocks).

This patch fixes this by always using at least 16 MiB for
the arbiter brick.

Note this required to change two of the test cases
slightly: The sizes are actually in KB. The tests use
some fantasy numbers like 800 KiB brick sizes that now
contradict for some checks the new mode of having a
minimum arbiter brick size of 16 MiB.

Fixes: #1181

Signed-off-by: Michael Adam <obnox@redhat.com>