apps: fix arbiter placer so that nodes do not get reused in a brick set
Due to changes introduced in a previous arbiter fix the function
that returns only "real" bricks from a brick set was not
using the correct indexes for the bricks in the brick set and
thus the check in the code for protecting against re-use of a node
within a single brick set was incorrect.
This fix returns the iteration to the previous approach and
checks for "nil bricks" in the loop itself.

Signed-off-by: John Mulligan <jmulligan@redhat.com>