executors: do not create backups in /etc/lvm/archive by default
With the automation that Heketi provides and the redundancy that Gluster
offers there is no need for the LVM-metadata backups under
/etc/lvm/archive. It is not expected that users have a need for these
archives, and they can take up a considerable amount space on the /
filesystem in active environments.

The heketi.json used by the service daemon can be configured to create
the LVM-metadata archives by setting "backup_lvm_metadata" to true. This
option defaults to false if not set.

BUG: https://bugzilla.redhat.com/1561680
Signed-off-by: Niels de Vos <ndevos@redhat.com>