executors: treat failure to unmount as a real error
Heketi has been ignoring errors from umount of bricks (and
failures to edit fstab) and assuming that a Brick destroy
always works. Now that we have the pending operations framework
we actually want to keep track of failures to delete so
that heketi (and the user) knows what was supposed to be
deleted but wasn't, allowing for targeted cleanups
and debugging.

Signed-off-by: John Mulligan <jmulligan@redhat.com>