db-upgrade: reduce complexity in addVolumeIdInBrickEntry()
Instead looping over
clusters -> volumes in the cluster-> bricks in the volume

loop over
volumes -> bricks in the volume

This reduces the complexity, while at the same
time preventing startup from failing if cluster's
volume-list has a volume-id which does not have
a volume entry in the DB any more.

This patch is best viewn with "git show -w"
(ignore white space changes).

Signed-off-by: Michael Adam <obnox@redhat.com>