functional tests: support custom methods for running gluster
This change slightly restructures the lib.sh functions and allows
for environment vars to be used to provide custom methods for
starting and stopping the gluster nodes.

The idea here is to move towards abstracting away the use of
vagrant and making it easy to try different approaches to
running the gluster test nodes.

Signed-off-by: John Mulligan <jmulligan@redhat.com>