apps: treat cleanup and remove cases for block hosting vols different
Unfortunately, block hosting volumes were structured different from
regular volumes in where the volume was mostly configured (pre-exec
vs. post-exec). With the introduction of operations this lead to
some incorrect code reuse in the delete and rollback cases in that
the create case does not update the block hosting volume size
immediately (this is a bug too, IMO) but only in finalize.
Thus trying to give space back to the block hosting volume in
rollback was incorrect behavior.
This somewhat hacky approach tries to fix this issue minimally
by not adjusting the block hosting volume size on rollback.

Future cleanups of the structure of block volume creation
should try to sort out what should be set before the exec, but
that's for another PR.

Signed-off-by: John Mulligan <jmulligan@redhat.com>