executors: delete path needs to check raw errors from lvm commands
When this function was refactored in preparation for robust delete
behavior the "friendly" error message was moved into the utility
function used for counting the number of lvs in a thin pool.
However, the errors from this function need to be checked to see
if lvm errors out because the thin pool is missing vs. another
reason. Changing the raw lvm error to a friendly error in this
case is wrong. Move the friendly error back to the upper function
call.

Fixes issue #1242

Signed-off-by: John Mulligan <jmulligan@redhat.com>