client api: combine retry behavior control under one options type
Combine the behavioral control over the client in a single
ClientOptions type. This type is able to regulate all
current and future core behaviors of the heketi client.
From this point on we should not need to keep adding
extra New* functions, simply extend the public attributes
of the ClientOptions type.

ClientOptions is responsible for controlling the enabling
of retries, the number of retries, and how much to delay between
retries.

Signed-off-by: John Mulligan <jmulligan@redhat.com>

xxx: make everything configurable

Signed-off-by: John Mulligan <jmulligan@redhat.com>