apps: add client throttling based on operations in-flight
Use the OpCounter type to throttle clients by rejecting new
operation when there the server is already busy processing
existing operations.
All operations that use the AsyncHttpOperation function will
be throttled as well as node and device removal
(set state when new state is failed). These are the activities
that most occupy the server.

Signed-off-by: John Mulligan <jmulligan@redhat.com>