utils: add shuffle methods to the util functions
The shuffle methods are useful for pseudo-randomly shuffling
an array in order to avoid picking the same N items every
time a particular list of elements is iterated over.

This is based on the implementation in Go 1.10 but since
Heketi supports older versions of Go we use a our own
internal one until we drop support for anything before Go 1.10.

Signed-off-by: John Mulligan <jmulligan@redhat.com>