apps:glusterfs update volume hosts info during node removal
during the node removal in heketi, we replace
the brick from removal node to another node,
during this process we need to update the
volume info such as Hosts (which may contain
the removal node hostname). and we need to
update the backup-volfile-servers and volume
mountpoint  (this also may contain the hostname)

if it's a block hosting volume, we need to
fetch all the block volumes belongs to it,
and update the hosts in those volumes, so that
next block volume creation fetches the new
hosts from the heketi cluster.

Signed-off-by: Madhu Rajanna <mrajanna@redhat.com>