apps: shuffle ha nodes to avoid relying on same nodes
When the number of nodes available is greater than the HA count
we want to select a subset of these nodes for use with gluster
block ha support. Previous changes to the code try to select
the ha-count number of healthy nodes from this set but if the
healthy node set was unchanging it would always pick the exact
same nodes.
This change shuffles the list of nodes that is tested for
liveness so different block volumes are distributed across
different nodes.

Signed-off-by: Sven Anderson <sven@redhat.com>