apps:glusterfs: reserve 1% of raw capacity in FreeSpace tracking
A block-hosting volume created with raw size X can not host block files
of accumulated size X because the file system uses a certain small
amount of the raw size. But the FreeSpace tracking does currently not
take this into account. Hence trying to create a block volume of size X
with FreeSize X available on the block hosting volume will fail.

This path fixes heketi to always track only 99% of the raw space of a
newly created (or expanded) block hosting volume as the available
FreeSize. Thus the announced free space will always be able to
host content of that size.

Signed-off-by: Michael Adam <obnox@redhat.com>