apps:glusterfs update volume hosts info during node deletion
during the node delete in heketi, we remove node from
heketi management,volume info such as Hosts (which may
contain the deleting node hostname). and we need to
update the backup-volfile-servers and volume mountpoint
(this also may contain the deleting node hostname)

if it's a block hosting volume, we need to
fetch all the block volumes belongs to it,
and update the hosts in those volumes, so that
next block volume creation fetches the new
hosts from the heketi cluster.

Signed-off-by: Madhu Rajanna <mrajanna@redhat.com>