executors: create an error injection executor
The error injection executor is an executor that intercepts activities
in a normal executor, such as the ssh executor and can produce
simulated error conditions at the executor interface level or before or
after running a command on a node.

Currently, the error inject executor can be configured to produce
errors (or fake results) based on a list of "hooks" that compare
commands (or commands & results) to regexes to produce some result
such as a normal error condition or a panic.

Signed-off-by: John Mulligan <jmulligan@redhat.com>