functional tests: add an env var that skips starting server
Add an environment variable (HEKETI_TEST_SERVER) that when set to "no"
will prevent the functions in lib.sh from starting the heketi server.
This allows the use of the lib.sh functions for setting up the
environment but allows the test cases to control the starting and
stopping of the server. This will be used to write tests that exercise
the error injection executor.

Signed-off-by: John Mulligan <jmulligan@redhat.com>