middleware: implement new claims type and validation function
Due to inability of jwt library to handle clock skew, we implement a
derived type of claims that does validation of "iat" claim with leeway.
As a side effect you see the following changes:
1. required_claims is embedded in validation function and missing
claims error is replaced with unauthorized
2. As claims type has changed from map to struct, claims are now
accessed as members and not indexed.

The leeway for "iat" is set to 120 seconds by default. It can only be
changed through env variable: HEKETI_JWT_IAT_LEEWAY_SECONDS.

Signed-off-by: Raghavendra Talur <rtalur@redhat.com>