apps: change Modify(Free|Reserved)Size function to return error
Change the VolumeEntry's ModifyFreeSize and ModifyReservedSize
functions to return errors instead of panicing. Also plumb
the error handling up to the callers of these functions.
This will prevent wrong values from being saved to the db
without killing the server. The admin of the server can then
stop the server at a later time and repair the db as needed.

Signed-off-by: John Mulligan <jmulligan@redhat.com>