apps: add db upgrade func to correct bad block hosting free sizes
Add a function that runs during db upgrade to find and try to correct
any bad block hosting volume free size counts.
This function will skip updating the free size to something outside
the valid values, so if the size is bad but the correction makes
no sense, it will leave things as is.

Signed-off-by: John Mulligan <jmulligan@redhat.com>