executors: format arbiter bricks differently
Add support to the brick request to tell the executor that a brick
file system should be formatted in one fashion or another. Define
a normal and arbiter formatting type and use the arbiter
formatting type to create a file system that allows for more
inodes.

Pair-Programmed-With: John Mulligan <jmulligan@redhat.com>
Signed-off-by: Raghavendra Talur <rtalur@redhat.com>
Signed-off-by: John Mulligan <jmulligan@redhat.com>