apps:glusterfs: reserve 2% of raw capacity in FreeSpace tracking
For larger block volumes, reserving 1% of raw capacity is
not enough, as the block volume delete operation in
gluster block also requires some memory to store its
metadata during deletion of block volume.

If we reserve 2% of raw capacity, we can create and delete
block volumes of larger size. tested with 10GB, 50GB,
100GB, 200GB, 500GB and 1TB) of block volumes.

Signed-off-by: Madhu Rajanna <mrajanna@redhat.com>