docker: load an initial topology if HEKETI_TOPOLOGY_FILE is set
By providing the environment variable HEKETI_TOPOLOGY_FILE with a
filename that is available in the container, the database gets populated
with this topology file.

In Kubernetes environments the HEKETI_TOPOLOGY_FILE can be provided as a
ConfigMap or Secret and mounted as a volume in the pods specification.
With this change it is now possible to deploy Heketi with a .yaml file
that contains the topology.json and the Heketi Deployment.

Signed-off-by: Niels de Vos <ndevos@redhat.com>