makefile: install binaries and container files
This change restructures how some of the container related files
are delivered. It adds a 'make install' target that installs
the heketi and heketi-cli binaries along with the manpage and
some supporting files. The fromsource container builds use these
files rather than pull them directly through the Dockerfile.
The end goal is to unify the heketi sources so that all sorts of
container builds can use the same files.

Signed-off-by: John Mulligan <jmulligan@redhat.com>