executors: prevent pvcreate/vgcreate/lvcreate from displaying an interactive prompt
In certain circumstances LVM commands can prompt the user for a
confirmation. Users will not be able to answer the prompt in case Heketi
executes the commands. By passing the `-qq` parameter to the LVM
commands, the default answer for the prompt will be no and the command
exits with a failure.

Fixes: gluster/gluster-kubernetes#497
Reported-bu: Ulf Lilleengen <lulf@redhat.com>
Signed-off-by: Niels de Vos <ndevos@redhat.com>