api: define a new type to restrict usage of block hosting volumes
A new type and corresponding property on the volume's BlockInfo
struct will be used to restrict if new block volumes can be
placed on a block hosting volume.

This can be used by admins for migrating off of one block hosting
volume. It will also be used by the system to mark "sketchy"
block hosting volumes post-upgrade.

Signed-off-by: John Mulligan <jmulligan@redhat.com>