apps: add an upgrade function that tries to update reserved size on BHVs
Add a new upgrade function that attempts to update the reserved
size on all block hosting volumes in the db. If the expected amount
of reserved size can't be taken the system will set the
"locked-by-update" restriction on the BHV. This restriction
can by cleared by the admin, but only if certain preconditions are
met (see change that adds operation).
If the values seen on BHVs are not sane the restriction will also be
set.

Signed-off-by: John Mulligan <jmulligan@redhat.com>