apps: add a test for insufficient hosts on HA count = 3
Add a test case that checks that the system fails as expected
if there are too few nodes working when an ha count of 3 is
given. Also checks the recent change to teat a does-not-exist
error from gluster block as a good exit condition for
block volume delete. This tests that when the exec step fails
due to ha count issues, that the overall operation is cleaned
up.

Signed-off-by: John Mulligan <jmulligan@redhat.com>