kubeexec: match the stderr output of kubeexec to that of sshexec
Kubeexec was adding extra text to stderr before sending it back to
caller. In cases where the stderr text has to be parsed it was leading
to parse errors.

It is sufficient to log the details of pod where the commands were
executed and such. Let the error string be passed as-is to the caller.

Signed-off-by: Raghavendra Talur <rtalur@redhat.com>