executors: use new method of gathering results in block-volume-delete
Block volume delete method was pretty hacky and relied on the shell
command (on the remote node) duplicating the results on stdout and stderr
to work. This change bases the function on the newer more comprehensive
Result object so that no shell hacks need to be done on the remote
side. At the same time, this change reorganizes the function a bit
to better use the Result type and for clarity.

Signed-off-by: John Mulligan <jmulligan@redhat.com>