remoteexec: create targets for resources w/in the k8s cluster
Create a family of target types that are to be used for
determining where commands can be run the k8s cluster. You start
with the type closest to the data you have and then resolve
that to a type that is closer to the container. For example:
target label -> target pod -> target container.

Signed-off-by: John Mulligan <jmulligan@redhat.com>