container: add tcp-user-timeout in default option
As identified in glusterfs patch[1], we need option
'server.user-tcp-timeout' to be set on gluster-block volume to mitigate
some of the 'hung status' like behavior when a node goes down. It is
also required for file volumes but the impact is not as critical. Adding
the option in heketi's default pre_request_volume_options will apply it
for all new volumes created by heketi.

For existing volumes, users can query the volume list and perform a
gluster volume set.

[1] https://review.gluster.org/#/c/glusterfs/+/21170/

Signed-off-by: Raghavendra Talur <rtalur@redhat.com>