container: allow to use current working repository for "fromsource" image
This builds makes it possible to build container images from the
currently checked-out sources. The extras/docker/fromsource/Dockerfile
configuration has a hardcoded git repository and branch. This is not
very flexible.

The following command will build the image, run it from the root of the
repository:

  $ buildah bud --network=host --file=extras/docker/fromsource/Dockerfile .

The old way of building is still supported. This makes the Dockerfile
look a little hacky, There are checks for missing files, and depending
of this the 'legacy' way of building is done, or the modern one. This is
split in separate RUN commands to enhance readability and maintenance.

With this change, a multi-stage build is introduced as well. This makes
sure the final container is clean and updated to the latest Fedora
version.

Signed-off-by: Niels de Vos <ndevos@redhat.com>