apps: protect against old state in block volume delete ops
Unlike volume creates, volume deletes require the system to check
the state of the pre-existing objects in the db. Thus the Build
step of the delete operations must update the state of the volume
in the db and not trust the state of the block volume entries
initially passed to the operation are up-to-date.

Signed-off-by: John Mulligan <jmulligan@redhat.com>