reintroduce rest framework into source tree
This change takes the source files for the formerly external
"heketi/rest" package and makes it a server utility library
(under "heketi/heketi/server/rest") with a license updated
to match the other server components.

The imports of the source files are adjusted to match the
internal "pkg/" utility packages, rather than the external
"heketi/utils" package.

This change is part of the effort to standardize and unify
licensing of the heketi packages as noted in issue #1279.
With this change the heketi server always uses the updated
in-tree utility functions.

Signed-off-by: John Mulligan <jmulligan@redhat.com>