client: convert topology info to use a template rather than printfs
Instead of using a complex chain of printf calls, change the topology
info command to use a template. This template matches the existing
content of topology info minus a few extra blank lines that
were left in because it made the template more readable and didn't
impact the readability of topology info.

Using a template ought to make future changes to the topology info
output easier and less brittle.

Signed-off-by: John Mulligan <jmulligan@redhat.com>